# swagger/api/v1/product.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/v1/product.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "ProductService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/products": {
      "get": {
        "summary": "Get all products",
        "operationId": "ProductService_GetProducts",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetProductsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "ProductService"
        ]
      },
      "post": {
        "summary": "Create a new product",
        "operationId": "ProductService_CreateProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateProductResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "product",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1Product"
            }
          }
        ],
        "tags": [
          "ProductService"
        ]
      }
    },
    "/v1/products/{id}": {
      "get": {
        "summary": "Get product by ID",
        "operationId": "ProductService_GetProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetProductResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "ProductService"
        ]
      },
      "delete": {
        "summary": "Delete a product",
        "operationId": "ProductService_DeleteProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteProductResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "ProductService"
        ]
      }
    },
    "/v1/products/{product.id}": {
      "put": {
        "summary": "Update an existing product",
        "operationId": "ProductService_UpdateProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateProductResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "product.id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "product",
            "description": "Product represents a product entity",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "colorId": {
                  "type": "string",
                  "format": "uint64"
                },
                "sizeId": {
                  "type": "string",
                  "format": "uint64"
                },
                "quantity": {
                  "type": "string",
                  "format": "uint64"
                },
                "available": {
                  "type": "string",
                  "format": "uint64"
                },
                "parentId": {
                  "type": "string",
                  "format": "uint64"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "deletedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "title": "Product represents a product entity"
            }
          }
        ],
        "tags": [
          "ProductService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CreateProductResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/v1Data"
        }
      },
      "title": "Response message for CreateProduct"
    },
    "v1Data": {
      "type": "object",
      "properties": {
        "products": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Product"
          }
        }
      }
    },
    "v1DeleteProductResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/v1Data"
        }
      },
      "title": "Response message for DeleteProduct"
    },
    "v1GetProductResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/v1Data"
        }
      },
      "title": "Response message for GetProduct"
    },
    "v1GetProductsResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/v1Data"
        }
      },
      "title": "Response message for GetProducts"
    },
    "v1Product": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "name": {
          "type": "string"
        },
        "colorId": {
          "type": "string",
          "format": "uint64"
        },
        "sizeId": {
          "type": "string",
          "format": "uint64"
        },
        "quantity": {
          "type": "string",
          "format": "uint64"
        },
        "available": {
          "type": "string",
          "format": "uint64"
        },
        "parentId": {
          "type": "string",
          "format": "uint64"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "deletedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "Product represents a product entity"
    },
    "v1UpdateProductResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/v1Data"
        }
      },
      "title": "Response message for UpdateProduct"
    }
  }
}